scripts/requirements.txt
pyyaml
inquirer
argparse